% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{assert_dependencies}
\alias{assert_dependencies}
\title{Assert that required packages are available}
\usage{
assert_dependencies(local, download)
}
\arguments{
\item{local}{Whether local Excel saving is requested}

\item{download}{Whether downloading is requested}
}
\description{
Assert that required packages are available
}
\keyword{internal}
