% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbarbGrob.R
\name{create_errorbarb}
\alias{create_errorbarb}
\title{Create Coordinates for Error Bars with Custom Tips}
\usage{
create_errorbarb(x, y, height, width, errorbar_tip_size)
}
\arguments{
\item{x}{Numeric value representing the x-coordinate of the center of the error bar}

\item{y}{Numeric value representing the y-coordinate of the center of the error bar}

\item{height}{Numeric value representing the height of the error bar}

\item{width}{Numeric value representing the width of the error bar}

\item{errorbar_tip_size}{Numeric value representing the size of the tips of the error bar}
}
\value{
A data frame containing the coordinates of the error bar segments
}
\description{
This function generates the coordinates required to draw custom error bars with specified height, width, and tip size.
}
\examples{
errorbar_coords <- create_errorbarb(0.5, 0.5, 0.1, 0.05, 0.02)
print(errorbar_coords)
}
