% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_reference_box.R
\name{draw_reference_box}
\alias{draw_reference_box}
\title{Draw a Reference Box on a ggplot}
\usage{
draw_reference_box(
  p,
  data,
  x,
  y,
  fill = "white",
  colour = "#028760",
  fun.errorbar = "sd",
  na.rm = FALSE,
  ...
)
}
\arguments{
\item{p}{A ggplot object to which the reference box will be added}

\item{data}{A data frame containing the data}

\item{x}{A character string specifying the name of the x variable}

\item{y}{A character string specifying the name of the y variable}

\item{fill}{The fill color of the reference box. Defaults to "white"}

\item{colour}{The border color of the reference box. Defaults to "#028760"}

\item{fun.errorbar}{A character string indicating the method to calculate the error bar. It can be either "sd" for standard deviation or "se" for standard error. Defaults to "sd"}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}

\item{...}{Other arguments passed on to \code{geom_rect()}}
}
\value{
A ggplot object with the reference box added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because more flexibile function \code{geom_errorbox()} is added.
}
\details{
This function adds a reference box to a ggplot, indicating the mean and error range for the specified x and y variables.
}
