% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbarbGrob.R
\name{errorbarbGrob}
\alias{errorbarbGrob}
\title{Create an Error Bar Grob}
\usage{
errorbarbGrob(
  x = grid::unit(0, "npc"),
  y = grid::unit(0, "npc"),
  fun.errorbar = "sd",
  na.rm = FALSE,
  errorbar_tip_size = grid::unit(0.1, "npc"),
  default.units = "npc",
  name = NULL,
  gp = grid::gpar(),
  vp = NULL
)
}
\arguments{
\item{x}{A numeric or unit object specifying the x-coordinates of the error bars. Defaults to \code{unit(0, "npc")}}

\item{y}{A numeric or unit object specifying the y-coordinates of the error bars. Defaults to \code{unit(0, "npc")}}

\item{fun.errorbar}{A character string indicating the method to calculate the error bar. It can be either "sd" for standard deviation or "se" for standard error. Defaults to "sd"}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}

\item{errorbar_tip_size}{A unit object specifying the size of the error bar tips. Defaults to \code{unit(0.1, "npc")}}

\item{default.units}{A character string indicating the default units to use if \code{x}, \code{y}, or \code{errorbar_tip_size} are not unit objects. Defaults to "npc"}

\item{name}{A character string specifying a name for the grob}

\item{gp}{An object of class \code{gpar}, typically the output from a call to the \code{gpar()} function, used to specify graphical parameters}

\item{vp}{A \code{viewport} object (or NULL)}
}
\value{
A gTree object of class "errorbarb"
}
\description{
This function creates a graphical object (grob) representing error bars for use in grid graphics.
}
\examples{
library(grid)

df <- data.frame(
  x = c(.4, .1),
  y = c(.1, .5)
)
errorbar_grob <- errorbarbGrob(
  x = df$x, y = df$y,
  fun.errorbar = "sd",
  errorbar_tip_size = unit(5, "cm")
)
grid.draw(errorbar_grob)
}
