% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_errorbarb.R
\name{Geomerrorbarb}
\alias{Geomerrorbarb}
\title{Geomerrorbarb ggproto object}
\description{
Internal ggproto backing \code{\link[=geom_errorbarb]{geom_errorbarb()}}. Most users should call
\code{geom_errorbarb()} directly.
}
\concept{ggplot2 geoms}
