% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorbarbGrob.R
\name{makeContent.errorbarb}
\alias{makeContent.errorbarb}
\title{Create Content for Error Bar Grob}
\usage{
\method{makeContent}{errorbarb}(grob)
}
\arguments{
\item{grob}{An object of class "errorbarb"}
}
\value{
A gTree object with the children set to the calculated error bars
}
\description{
This function defines the content of the error bar grob, calculating the positions and dimensions of the error bars.
}
