% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-errorbox.R
\name{StatErrorbox}
\alias{StatErrorbox}
\title{ggproto object for error box statistics}
\format{
ggproto class
}
\usage{
StatErrorbox
}
\description{
ggproto object for error box statistics
}
\keyword{internal}
