% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{assert_parameters}
\alias{assert_parameters}
\title{Validate function parameters}
\usage{
assert_parameters(
  .data,
  sheet_names,
  name,
  local,
  download,
  overwrite,
  filter = TRUE,
  freeze_first_row = TRUE,
  auto_width = TRUE
)
}
\arguments{
\item{.data}{List of dataframes}

\item{sheet_names}{Vector of sheet names}

\item{name}{Spreadsheet/file name}

\item{local}{Local Excel option}

\item{download}{Download option}

\item{overwrite}{Overwrite option; a single logical value}

\item{filter}{Filter option}

\item{freeze_first_row}{Freeze first row option}

\item{auto_width}{Auto-width option}
}
\description{
Validate function parameters
}
\keyword{internal}
