% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assert_parameters}
\alias{assert_parameters}
\title{Validate function parameters}
\usage{
assert_parameters(.data, sheet_names, name, local, download)
}
\arguments{
\item{.data}{List of dataframes}

\item{sheet_names}{Vector of sheet names}

\item{name}{Spreadsheet/file name}

\item{local}{Local Excel option}

\item{download}{Download option}
}
\description{
Validate function parameters
}
\keyword{internal}
