% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_error}
\alias{calc_error}
\title{Calculate SD/SE for Errorbar}
\usage{
calc_error(x, fun.errorbar = "sd", na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{fun.errorbar}{A character string indicating the method to calculate the error bar. It can be either "sd" for standard deviation or "se" for standard error. Defaults to "sd"}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}
}
\value{
The calculated error bar (either standard deviation or standard error) of the input vector
}
\description{
This function calculates either the standard deviation or the standard error of a numeric vector, depending on the specified method.
}
\examples{
calc_error(c(1, 2, 3, 4, 5))
calc_error(c(1, 2, 3, 4, 5, NA), fun.errorbar = "se", na.rm = TRUE)
}
