% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_error}
\alias{calc_error}
\title{Calculate SD/SE for Errorbar}
\usage{
calc_error(x, fun.errorbar = "sd", na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{fun.errorbar}{A character string or function indicating how to calculate the error bar. Character options are "sd" (standard deviation), "se" (standard error) and "ci" (95\% confidence interval). A custom function can also be supplied. Defaults to "sd"}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}
}
\value{
The calculated error bar (either standard deviation or standard error) of the input vector
}
\description{
This function calculates an error statistic for a numeric vector using the specified method.
}
\examples{
calc_error(c(1, 2, 3, 4, 5))
calc_error(c(1, 2, 3, 4, 5, NA), fun.errorbar = "se", na.rm = TRUE)
calc_error(c(1, 2, 3, 4, 5), fun.errorbar = "ci")
calc_error(c(1, 2, 3, 4, 5), fun.errorbar = function(z) max(z) - min(z))
}
