% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{download_google_sheet}
\alias{download_google_sheet}
\title{Download a Google Spreadsheet as Excel}
\usage{
download_google_sheet(spreadsheet_name, file_path)
}
\arguments{
\item{spreadsheet_name}{Name of the spreadsheet to download}

\item{file_path}{Path to save the Excel file}
}
\value{
The file path if successful, NULL otherwise
}
\description{
Download a Google Spreadsheet as Excel
}
\keyword{internal}
