% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fix_limit}
\alias{fix_limit}
\title{Fix Aspect Ratio of ggplot Based on Plot Limits}
\usage{
fix_limit(.plot, .ratio, .clip = "off")
}
\arguments{
\item{.plot}{A ggplot object}

\item{.ratio}{Numeric value to modify the calculated aspect ratio}

\item{.clip}{Character string specifying the clipping behavior ("off" by default)
See \code{\link[ggplot2]{coord_fixed}} for more details}
}
\value{
A modified ggplot object with adjusted aspect ratio
}
\description{
This function adjusts the aspect ratio of a ggplot object by calculating the ratio
based on the current plot limits and a desired ratio modifier.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point()

# Adjust the aspect ratio
fix_limit(p, .ratio = 1)

}
