% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_errorbox.R
\name{geom_errorbox}
\alias{geom_errorbox}
\title{Add error boxes to a plot}
\usage{
geom_errorbox(
  mapping = NULL,
  data = NULL,
  stat = "errorbox",
  position = "identity",
  ...,
  fun.errorbar = "sd",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings, usually created with \code{aes()}}

\item{data}{The data to be displayed. If NULL, the default, the data is inherited from the plot data}

\item{stat}{The statistical transformation to use on the data}

\item{position}{Position adjustment}

\item{...}{Other arguments passed to \code{GeomRect}}

\item{fun.errorbar}{Error calculation method ("sd", "se", "ci" or a custom function)}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If FALSE, overrides the default aesthetics}
}
\value{
A ggplot2 layer
}
\description{
This function draws a rectangular error box centered at the mean of x and y coordinates,
with dimensions determined by a specified error calculation method (standard deviation,
standard error, confidence interval, etc.).
}
\details{
The error box is drawn as a rectangle spanning from xmin to xmax and ymin to ymax,
where these boundaries are calculated based on the specified error method.
The box is centered at the mean values of x and y coordinates.

Error calculation methods:
\itemize{
\item "sd": Standard deviation
\item "se": Standard error (standard deviation divided by square root of the number of observations)
\item "ci": 95\\% confidence interval
\item custom function: You can provide your own error calculation function
}
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg))
p + geom_point() +
  geom_errorbox(fill = "white", color = "blue", alpha = 0.3)

# Using standard error
p + geom_point() +
  geom_errorbox(fill = "white", color = "red", alpha = 0.2, fun.errorbar = "sd")

# Using confidence interval
p + geom_point() +
  geom_errorbox(fill = "white", color = "green", alpha = 0.2, fun.errorbar = "se")

# Error boxes by group
p <- ggplot(mtcars, aes(wt, mpg, color = factor(cyl)))
p + geom_point() +
  geom_errorbox(aes(fill = factor(cyl)), alpha = 0.2)
}

}
