% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{se}
\alias{se}
\title{Calculate Standard Error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to FALSE}
}
\value{
The standard error of the input vector
}
\description{
This function calculates the standard error of a numeric vector.
}
\examples{
se(c(1, 2, 3, 4, 5))
se(c(1, 2, 3, 4, 5, NA), na.rm = TRUE)
}
