% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_google_sheets}
\alias{write_google_sheets}
\title{Write dataframes to Google Sheets}
\usage{
write_google_sheets(data_list, sheet_names, spreadsheet_name)
}
\arguments{
\item{data_list}{List of dataframes}

\item{sheet_names}{Vector of sheet names}

\item{spreadsheet_name}{Name of the spreadsheet}
}
\value{
A tibble with spreadsheet ID and URL
}
\description{
Write dataframes to Google Sheets
}
\keyword{internal}
