% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_local_excel}
\alias{write_local_excel}
\title{Write dataframes to a local Excel file}
\usage{
write_local_excel(data_list, sheet_names, file_path)
}
\arguments{
\item{data_list}{List of dataframes}

\item{sheet_names}{Vector of sheet names}

\item{file_path}{Path to save the Excel file}
}
\value{
A tibble with the file path
}
\description{
Write dataframes to a local Excel file
}
\keyword{internal}
